+++ e2e/fixtures/fixture-helper.js
@@ -7,7 +7,7 @@ import axios from 'axios';
 import path from 'path';
 import createStaticServer from '../create-static-server';
 import { getFixturesServerPort, getLocalTestDappPort } from './utils';
-import blacklistURLs from '../resources/test-parameters';
+import blacklistURLs from '../resources/blacklistURLs';
 
 export const DEFAULT_DAPP_SERVER_PORT = 8085;
