+++ e2e/helpers.js
@@ -4,7 +4,7 @@ import {
   getGanachePort,
   getLocalTestDappPort,
 } from './fixtures/utils';
-import blacklistURLs from './resources/test-parameters';
+import blacklistURLs from './resources/blacklistURLs';
 
 export default class TestHelpers {
   static async waitAndTap(elementId, timeout, index) {
