+++ e2e/resources/blacklistURLs.js
