+++ e2e/specs/accounts/create-wallet-account.spec.js
@@ -3,7 +3,7 @@ import { Regression } from '../../tags';
 import WalletView from '../../pages/WalletView';
 import { importWalletWithRecoveryPhrase } from '../../viewHelper';
 import AccountListView from '../../pages/AccountListView';
-import blacklistURLs from '../../resources/test-parameters';
+import blacklistURLs from '../../resources/blacklistURLs';
 
 describe(Regression('Create wallet account'), () => {
   beforeAll(async () => {
