+++ e2e/specs/networks/connect-test-network.spec.js
@@ -3,7 +3,7 @@ import { importWalletWithRecoveryPhrase } from '../../viewHelper';
 import WalletView from '../../pages/WalletView';
 import NetworkListModal from '../../pages/modals/NetworkListModal';
 import NetworkEducationModal from '../../pages/modals/NetworkEducationModal';
-import blacklistURLs from '../../resources/test-parameters';
+import blacklistURLs from '../../resources/blacklistURLs';
 
 const GORELI = 'Goerli Test Network';
 const ETHEREUM = 'Ethereum Main Network';
